requirements_app.txt
python-telegram-bot==13.15
pyyaml>=6.0.0
deep-translator>=1.9.2
omegaconf==2.3.0
llama-cpp-python>=0.1.83
num2words>=0.5.12
transformers>=4.33.0
Pillow>=10.0.0
torch>=2.0.1
backoff>=2.2.1
langchain>=0.0.286
requests>=2.31.0
urllib3>=2.0.4
python-dotenv==1.0.0